ai_backend/requirements.txt
Flask
Flask-Cors
tensorflow
numpy
Pillow
python-dotenv
openai